User score adjusting according to difficulty

+++ backend/CodeyBE.Contracts/DTOs/ExerciseDTO.cs
@@ -11,7 +11,7 @@ namespace CodeyBE.Contracts.DTOs
     {
         public int PrivateId { get; set; } = ex.PrivateId;
         public string Type { get; set; } = ex.Type;
-        public int Difficulty { get; set; } = ex.Difficulty;
+        public double Difficulty { get; set; } = ex.Difficulty;
         public string? Statement { get; set; } = ex.Statement;
         public string? SpecificTip { get; set; } = ex.SpecificTip;
     }
